+++ src/Curiosity.DAL.EF/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.1.1] - 2021-07-29
+
+### Added
+
+- Added pagination helper.
+
 ## [1.1.0] - 2021-03-29
 
 ### Added